% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTU.R
\name{scotch_gene}
\alias{scotch_gene}
\title{Differential gene expression analysis using Wilcoxon Rank Sum test}
\usage{
scotch_gene(X1_gene, X2_gene, epsilon = 0.01, ncores = 10)
}
\arguments{
\item{X1_gene}{cell(row) x gene(column) count matrix for cell population 1.}

\item{X2_gene}{cell(row) x gene(column) count matrix for cell population 2.}

\item{epsilon}{a small number to add on gene count matrix for computational purpose.}

\item{ncores}{number of cores to use for parallel computing}
}
\value{
result dataframe.pct1 and pct2: cell percentage expressing the gene; logFC: log fold change; p_gene: p-value for the test.
}
\description{
Differential gene expression analysis using Wilcoxon Rank Sum test
}
\examples{
df_gene = scotch_gene(sample8_CD4_gene, sample8_others_gene, epsilon=0.01,ncores=10)
}
