% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTU.R
\name{scotch_transcript}
\alias{scotch_transcript}
\title{Differential transcript usage analysis between two cell populations}
\usage{
scotch_transcript(
  gene_transcript_df1,
  gene_transcript_df2,
  X1_transcript,
  X2_transcript,
  ncores = 10,
  group_novel = TRUE
)
}
\arguments{
\item{gene_transcript_df1}{dataframe storing gene and transcript names for the transcript count matrix: X1_transcript.}

\item{gene_transcript_df2}{dataframe storing gene and transcript names for the transcript count matrix: X2_transcript.}

\item{X1_transcript}{cell(row) x transcript(column) count matrix for cell population 1.}

\item{X2_transcript}{cell(row) x transcript(column) count matrix for cell population 2.}

\item{ncores}{number of cores to use for parallel computing}

\item{group_novel}{whether to group all novel isoforms for DTU test}
}
\value{
result dataframe.
}
\description{
Differential transcript usage analysis between two cell populations
}
\examples{
df_transcript = scotch_transcript(gene_transcript_CD4_df,gene_transcript_CD8_df, sample8_CD4_transcript, sample8_CD8_transcript, 
ncores=10)
}
